fix: address PR review — retry logic, Y2038, DRY buffer pool, page size
- store_sent_message: use with_retry for SQLITE_BUSY resilience
- sent_messages.created_at: BigInt (i64) in schema to avoid Y2038 overflow
- Extract try_recycle_buffer() helper for 3 duplicated buffer pool blocks
- Guard coordination cache capacities with .max(1) to prevent zero-cap breakage
- rss_kib(): parse VmRSS from /proc/self/status instead of hardcoded page size

+++ storages/sqlite-storage/src/schema.rs
@@ -158,7 +158,7 @@ diesel::table! {
         message_id -> Text,
         payload -> Binary,
         device_id -> Integer,
-        created_at -> Integer,
+        created_at -> BigInt,
     }
 }
